sidehooks for drinkingID getter
