+++ app/assets/scripts/components/SettingsModal.js
@@ -3,8 +3,6 @@ import React, { useEffect } from "react"
 function SettingsModal(props) {
   return (
     <>
-      <div onClick={props.toggleSettings} className="modal-overlay"></div>
-
       <div className="modal-card"></div>
     </>
   )
