+++ app/assets/styles/modules/_settings-modal.css
@@ -1,18 +1,12 @@
-.modal-overlay {
-  position: absolute;
-  inset: 0;
-  background-color: rgba(173, 216, 230, 0.466);
-  z-index: 99;
-}
-
 .modal-card {
   position: absolute;
-  top: 20%;
-  right: 20%;
+  right: 30;
+  top: 30;
   background-color: #131a2a;
   width: 300px;
-  height: 300px;
+  height: 400px;
   z-index: 100;
   border: 1px solid $mainColor;
   border-radius: 20px;
+  box-shadow: 8px 15px 3px 10px rgba(0, 0, 0, 0.253);
 }
